pyproject.toml
[tool.poetry]
name = "loq"
version = "0.1.0"
description = "Social network for travelers"
authors = ["rusty <golovinartem2004@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
alembic = "^1.6.5"
Flask = "^2.0.1"
Flask-Login = "^0.5.0"
Flask-Migrate = "^3.0.1"
Flask-WTF = "^0.15.1"
Jinja2 = "^3.0.1"
psycopg2-binary = "^2.8.6"
SQLAlchemy-Utils = "^0.37.6"
WTForms-Alchemy = "^0.17.0"
WTForms-Components = "^0.10.5"
Flask-Security = "^3.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
